// src/app/common/enums.ts
export enum dificultad {
    superfacil = 2,
    facil = 4,
    medio = 5,
    dificil = 6,
    superdificil = 9
}

export enum nombresDificultad {
    superfacil = '🥒 Pickle Rick',
    facil='🌍 Morty',
    medio = '🍷 Beth',
    dificil = '🥼 Rick',
    superdificil= '💀 Super Morty'
}

// TIEMPO EN SEGUNDOS
export enum tiempoLimiteJuegos {
    superfacil = 45,
    facil = 60,
    medio = 90,
    dificil = 120,
    superdificil = 150
}

export enum tiempoPrevisualizacion {
    corto = 3,
    normal = 7,
    largo = 10
}